companion/requirements.txt
PyQt6>=6.6.0
watchdog>=3.0.0
keyboard>=0.13.5
imageio-ffmpeg>=0.4.9
psutil>=5.9.0
discord.py>=2.3.0

